pyproject.toml
[tool.poetry]
name = "kolomon-backend"
version = "0.1.0"
description = "A backend for Stari kolomoni project"
authors = ["Grimpy101, DefaultSimon"]

[tool.poetry.dependencies]
python = "^3.9"
drf-spectacular = "^0.17.3"
fastapi = {extras = ["all"], version = "^0.68.1"}
pytest = "^6.2.5"
SQLAlchemy = "^1.4.25"
psycopg2 = "^2.9.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
